+++ source/kCura.IntegrationPoints.Core/Factories/Implementations/ManagerFactory.cs
@@ -56,7 +56,7 @@ namespace kCura.IntegrationPoints.Core.Factories.Implementations
 			return new ErrorManager(CreateRepositoryFactory(contextContainer));
 		}
 
-		public IJobStopManager CreateJobStopManager(IContextContainer contextContainer, IJobService jobService, IJobHistoryService jobHistoryService, Guid jobIdentifier, int jobId)
+		public IJobStopManager CreateJobStopManager(IContextContainer contextContainer, IJobService jobService, IJobHistoryService jobHistoryService, Guid jobIdentifier, long jobId)
 		{
 			return new JobStopManager(jobService, jobHistoryService, jobIdentifier, jobId);
 		}
